+++ R/Brazil.R
@@ -19,8 +19,8 @@ Brazil <- R6::R6Class("Brazil",
   public = list(
 
     # Core Attributes
-    #' @field country name of country to fetch data for
-    country = "Brazil",
+    #' @field origin name of origin to fetch data for
+    origin = "Brazil",
     #' @field supported_levels A list of supported levels.
     supported_levels = list("1", "2"),
     #' @field supported_region_names A list of region names in order of level.