+++ R/SouthAfrica.R
@@ -16,8 +16,8 @@ SouthAfrica <- R6::R6Class("SouthAfrica",
   inherit = DataClass,
   public = list(
     # Core Attributes
-    #' @field country name of country to fetch data for
-    country = "South Africa",
+    #' @field origin name of origin to fetch data for
+    origin = "South Africa",
     #' @field supported_levels A list of supported levels.
     supported_levels = list("1"),
     #' @field supported_region_names A list of region names in order of level.