+++ R/WHO.R
@@ -16,8 +16,8 @@ WHO <- R6::R6Class("WHO",
   public = list(
 
     # Core Attributes
-    #' @field country name of country to fetch data for
-    country = "World Health Organisation (WHO)",
+    #' @field origin name of origin to fetch data for
+    origin = "World Health Organisation (WHO)",
     #' @field supported_levels A list of supported levels.
     supported_levels = list("1"),
     #' @field supported_region_names A list of region names in order of level.